Implement Getting Index & Save & Send to Group

+++ exchange/serializers.py
@@ -6,7 +6,7 @@ import datetime
 class IndexSerializer(serializers.ModelSerializer):
     value = serializers.FloatField(source="index", read_only=True)
     timestamp = serializers.SerializerMethodField(read_only=True)
-    datetime = serializers.IntegerField(write_only=True, )
+    datetime = serializers.FloatField(write_only=True, )
     index = serializers.FloatField(write_only=True, )
 
     def get_timestamp(self, obj: Index):
